requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
django-otp==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
pipenv==2020.6.2
psycopg2==2.7.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
twilio==6.44.0
urllib3==1.25.9
virtualenv==20.0.27
virtualenv-clone==0.5.4
whitenoise==5.1.0
zipp==3.1.0
